requirements.txt
numpy
scipy
matplotlib
Cython
hydra-core
joblib
setuptools
sklearn_crfsuite
pytorch-transformers
tensorflow
pyprind
seqeval
git+https://github.com/IINemo/libact.git@seq
git+https://github.com/IINemo/active_learning_toolbox.git@seq
git+https://github.com/IINemo/bert_sequence_tagger.git
flair
allennlp
nltk